+++ lib/host-sp-messages/Cargo.toml
@@ -19,4 +19,4 @@ zerocopy.workspace = true
 oxide-barcode.path = "../oxide-barcode"
 unwrap-lite.path = "../unwrap-lite"
 drv-i2c-types.path = "../../drv/i2c-types" # to implement From<ResponseCode>
-task-sensor-api.path = "../../task/sensor-api"
+task-sensor-types.path = "../../task/sensor-types"
