+++ task/sensor-api/Cargo.toml
@@ -10,9 +10,9 @@ num-traits.workspace = true
 serde.workspace = true
 zerocopy.workspace = true
 
-derive-idol-err.path = "../../lib/derive-idol-err"
 drv-i2c-api.path = "../../drv/i2c-api"
 userlib.path = "../../sys/userlib"
+task-sensor-types.path = "../sensor-types"
 
 # This section is here to discourage RLS/rust-analyzer from doing test builds,
 # since test builds don't work for cross compilation.
